telegram mining work

+++ helpers.go
@@ -518,7 +518,7 @@ func telegramMine(code string, tid int64) string {
 		logTelegram(err.Error())
 	}
 
-	mNotCode := fmt.Sprintf("This code is not valid, it should be 3 numbers.\n\nYou can see the daily mining code <a href=\"https://t.me/AnoteToday/%d\">here</a>.", adnum.(int))
+	mNotCode := fmt.Sprintf("This code is not valid, it should be 3 numbers.\n\nYou can see the daily mining code <a href=\"https://t.me/AnoteToday/%d\">here</a>.", adnum.(int64))
 
 	if len(code) != 3 {
 		return mNotCode
